fix: add dashboard view for account with cards
