+++ packages/blockchain-wallet-v4-frontend/src/data/components/debitCard/selectors.ts
@@ -1,6 +1,5 @@
 import { isEmpty } from 'ramda'
 
-import { selectors } from 'data'
 import { RootState } from 'data/rootReducer'
 
 export const getProducts = (state: RootState) => state.components.debitCard.products
@@ -11,3 +10,5 @@ export const isDebitCardModuleEnabledForAccount = (state: RootState): boolean =>
   // If feature flag is disabled then it will only be the initial state in products
   return !isEmpty(getProducts(state))
 }
+
+export const getCards = (state: RootState) => state.components.debitCard.cards
