+++ packages/blockchain-wallet-v4-frontend/src/scenes/DebitCard/CardDashboard/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+
+// TODO: Will continue this in CI-76
+const CardDashboard = () => <>Card Dashboard</>
+
+export default CardDashboard
