+++ packages/blockchain-wallet-v4-frontend/src/scenes/DebitCard/template.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { Button, Icon, Image } from 'blockchain-info-components'
+import { Icon } from 'blockchain-info-components'
 import {
   HeaderTextWrapper,
   IconBackground,
@@ -12,46 +12,11 @@ import {
 import { ModalName } from 'data/modals/types'
 
 import { Props } from '.'
-import { IntroCardWrapper, Wrapper } from './model'
+import CardDashboard from './CardDashboard'
+import CardOrder from './CardOrder'
+import { Wrapper } from './model'
 
-type ButtonProps = {
-  onClick: () => void
-}
-
-const OrderMyCardButton = ({ onClick }: ButtonProps) => (
-  <div style={{ marginTop: '32px' }}>
-    <Button
-      data-e2e='orderMyCard'
-      nature='primary'
-      height='48px'
-      width='327px'
-      margin='auto'
-      onClick={onClick}
-    >
-      <FormattedMessage id='scenes.debit_card.intro.order_my_card' defaultMessage='Order My Card' />
-    </Button>
-  </div>
-)
-
-const LinkHereButton = () => (
-  <div style={{ marginTop: '16px' }}>
-    <Button
-      data-e2e='linkCard'
-      nature='empty-blue'
-      height='48px'
-      width='327px'
-      margin='auto'
-      onClick={() => {}}
-    >
-      <FormattedMessage
-        id='scenes.debit_card.intro.already_have_a_card'
-        defaultMessage='Already Have A Card? Link It Here'
-      />
-    </Button>
-  </div>
-)
-
-const DebitCard = ({ modalActions }: Props) => {
+const DebitCard = ({ cards, modalActions }: Props) => {
   const handleOpenOrderMyCard = () =>
     modalActions.showModal(ModalName.ORDER_MY_CARD, { origin: 'DebitCardPage' })
 
@@ -73,28 +38,11 @@ const DebitCard = ({ modalActions }: Props) => {
           defaultMessage='Taking crypto into the physical world.'
         />
       </SceneSubHeaderText>
-      <IntroCardWrapper>
-        <Image name='intro-card' />
-
-        <SceneHeaderText>
-          <FormattedMessage
-            id='scenes.debit_card.intro.content.header'
-            defaultMessage='Your Gateway To The Blockchain Debit Card'
-          />
-        </SceneHeaderText>
-
-        <SceneSubHeaderText>
-          <FormattedMessage
-            id='scenes.debit_card.intro.content.subheader'
-            defaultMessage='A card that lets you spend and earn in crypto right from your Blockchain account.'
-          />
-        </SceneSubHeaderText>
-        <OrderMyCardButton onClick={handleOpenOrderMyCard} />
-        {/*
-          Waiting for UX designs
-          <LinkHereButton />
-        */}
-      </IntroCardWrapper>
+      {cards.length === 0 ? (
+        <CardOrder handleOpenOrderMyCard={handleOpenOrderMyCard} />
+      ) : (
+        <CardDashboard />
+      )}
     </Wrapper>
   )
 }
